Change Gradle build from multi-project build to composite build
A composite build is the more correct mechanism to use here, because
prob2-ui is an independent project, not a subproject of
EventBPacman-Plugin. This is also necessary for certain Gradle features
to behave correctly (e. g. wrapper configuration in the prob2-ui
build.gradle) and for when a built prob2-ui is published on Maven
Central at some point.