Change how the plugin updates the trace to improve performance
Previously every state change would update the currentTrace, which
would trigger a full UI update, even for intermediate states that the
user would never see in the end. This made the updates after movements
very slow.

Now the currentTrace is only updated once per movement, and all
internal methods receive the trace as an argument (and return a new
trace if necessary). This improves performance, at the cost of making
the code a bit more verbose in some places.